requirements.txt
python-pptx
beautifulsoup4
requests
